+++ CHANGELOG.md
@@ -29,7 +29,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - [NSEK-109](https://jira.itkdev.dk/browse/NSEK-84): Adding documentation about translations.
 - [NSEK-112](https://jira.itkdev.dk/browse/NSEK-112): Adding EasyAdmin bundle.
 - [NSEK-105](https://jira.itkdev.dk/browse/NSEK-105): Added CLI and Azure B2C login.
-- [TVIST-237](https://jira.itkdev.dk/browse/TVIST1-237): Logging documentation.
+- [TVIST1-237](https://jira.itkdev.dk/browse/TVIST1-237): Logging documentation.
+- [TVIST1-253](https://jira.itkdev.dk/browse/TVIST1-253): Added AbstractEntityListener, CaseListener,
+MunicipalityListener, BoardListener and their tests.
 
 ### Changed
 - [NSEK-138](https://jira.itkdev.dk/browse/NSEK-138): Updating Cypress to 6.5.0 so the same version is used project wide.